% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indmortality-package.R
\docType{data}
\name{aus}
\alias{aus}
\alias{nsw}
\alias{vic}
\alias{qld}
\alias{nt}
\alias{sa}
\alias{wa}
\title{Lists of demogdata objects for all Australian states, 2001-2007}
\description{
Objects: \code{aus}, \code{nsw}, \code{vic}, \code{sa}, \code{wa},
\code{nt}, and \code{qld}.
Note that there is insufficient data to produce similar objects for Tasmania and the ACT.
These are demogdata objects, based on backcast population estimates.
The rates are smoothed over age using \code{\link[demography]{smooth.demogdata}}.
For information about demogdata objects, see \code{\link[demography]{demogdata}}.
}
\details{
Demogdata objects constructed using \code{\link{make.demogdata}} using
2001-2010 data from \code{\link{ideaths}} and \code{\link{ipop}}.
}
\examples{

plot(aus,'male')

# To create similar objects using interpolated population,
# use, for example,
nsw2 <- make.demogdata(state="NSW", population="interpolated")

}
\keyword{datasets}
