% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtable.R
\name{deathstable}
\alias{deathstable}
\title{Function to make deaths tables for subsets of data.}
\usage{
deathstable(state = c("AUS", "ACT", "NSW", "NT", "QLD", "SA", "TAS", "VIC",
  "WA"), sex = c("female", "male", "total"), year = 2001:2010,
  upper.age = 115)
}
\arguments{
\item{state}{The function will return all deaths from the state indicated.}

\item{sex}{The function will return all deaths from the sex indicated.}

\item{year}{Which years to be counted. Default is all available years.}

\item{upper.age}{Deaths are cumulated for higher ages into this upper age
group.}
}
\value{
A numerical vector with death counts in each cell.
}
\description{
Returns vector of total deaths by single-year of age. 
Deaths are counted using the method adopted for the 
``Enhanced Mortality Database'', adjusting for the 
under-identification of Indigenous deaths on the 
National Death Index.
}
\examples{

tab1 <- deathstable(sex="female")
tab2 <- deathstable(state="VIC", cut="H", aveyear=TRUE)
tab3 <- deathstable()

}
\references{
Choi, C., Hyndman, R.J., Smith, L., and Zhao, K. (2010) \emph{An
enhanced mortality database for estimating indigenous life expectancy}.
Report for Australian Institute of Health and Welfare.
}
\author{
Rob J Hyndman <Rob.Hyndman@monash.edu>
}
