% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indmortality-package.R
\docType{package}
\name{indmortality-package}
\alias{indmortality-package}
\alias{indmortality}
\title{Australian indigenous mortality rates and life expectancies}
\description{
Package prepared for the Australian Institute of Health and Welfare to
compute life expectancy and other mortality-related information for the
Australian indigenous population using simulation and nonparametric
smoothing.
}
\details{
\tabular{ll}{ Package: \tab indmortality\cr Type: \tab Package\cr Version:
\tab 1.0\cr License: \tab GPL 2\cr }
}
\examples{

nsw.nmd <- make.demogdata(state="NSW", cut="M", sex="male", aveyear=TRUE)
plot(nsw.nmd)
life.expectancy.mean(nsw.nmd)
life.expectancy.ci(nsw.nmd)

}
\references{
Choi, C., Hyndman, R.J., Smith, L., and Zhao, K. (2010) \emph{An
enhanced mortality database for estimating indigenous life expectancy}.
Report for Australian Institute of Health and Welfare.
}
\seealso{
\code{\link[demography:demography-package]{demography}}
}
\author{
Rob J Hyndman <Rob.Hyndman@monash.edu>

Maintainer: Rob J Hyndman <Rob.Hyndman@monash.edu>
}
\keyword{package}
