% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life.expectancy.all.R
\name{life.expectancy.mean}
\alias{life.expectancy.mean}
\alias{life.expectancy.ci}
\title{Calculate life expectancies for both sexes and all cuts of the deaths
matrix.}
\usage{
life.expectancy.mean(object, ages = c(0, 20, 45, 65))
}
\arguments{
\item{object}{Output from \code{\link{make.demogdata}}.}

\item{ages}{Ages for which life expectancy is required.}

\item{nsim}{Number of simulations to use in the calculation.}

\item{level}{Confidence level for the intervals.}

\item{smooth}{If TRUE, smoothing will be used before computing life
expectancies.}

\item{sigma}{Value of sigma used in PES estimate for simulations.}
}
\value{
An array containing life expectancies for each year, age, and sex 
contained in \code{object}.
}
\description{
\code{life.expectancy.mean} calculates life expectancies and
\code{life.expectancy.ci} computes confidence intervals for life
expectancies. The functions take the output from
\code{\link{make.demogdata}} and returns life expectancy results for 
both sexes and all years for the specified ages.
}
\examples{

life.expectancy.mean(nsw)
life.expectancy.ci(nsw)

}
\references{
Choi, C., Hyndman, R.J., Smith, L., and Zhao, K. (2010) \emph{An
enhanced mortality database for estimating indigenous life expectancy}.
Report for Australian Institute of Health and Welfare.
}
\seealso{
\link[demography]{life.expectancy}
}
\author{
Rob J Hyndman <Rob.Hyndman@monash.edu>
}
