% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.demogdata.R
\name{make.demogdata}
\alias{make.demogdata}
\title{Make demogdata objects for particular states and years.}
\usage{
make.demogdata(state = c("AUS", "ACT", "NSW", "NT", "QLD", "SA", "TAS", "VIC",
  "WA"), sex = c("ALL", "female", "male", "total"), cut = c("ALL", "M", "H",
  "A", "N"), year = NULL, aveyear = FALSE, upper.age = 100,
  smooth = TRUE, deaths = indmortality::ideaths, pop = indmortality::ipop)
}
\arguments{
\item{state}{Character code indicating state or territory of Australia, or
"AUS" indicating the whole of Australia.}

\item{sex}{If "ALL", the function will return all sexes as part of each
demogdata object. Otherwise the function will return only the sex indicated.}

\item{cut}{Which cut of data to use. If "ALL", then a list of demogdata
objects is returned.}

\item{year}{Years to include in the demogdata object. The default is to
include all available years from the deaths matrix.}

\item{aveyear}{Logical value indicating whether results should be averaged
across years. Default is FALSE.}

\item{upper.age}{Upper age group, by default set to 100+.}

\item{smooth}{Logical value indicating whether mortality rates should be
smoothed using penalized regression splines. Default is TRUE.}

\item{deaths}{A matrix containing all unit death records in the form of
\code{\link{ideaths}}.}

\item{pop}{A list of list of matrices containing population numbers. This
should be in the form of \code{\link{ipop}}.}
}
\value{
A demogdata object (if \code{cut!="ALL"}) or a list of demogdata
objects (if \code{cut=="ALL"}). See \link[demography]{demogdata} for more
information about demogdata objects.
}
\description{
The function \code{make.demogdata} will return a demogdata object if
\code{cut!="ALL"}. Otherwise it returns a list of demogdata objects for all
cuts for a particular state.
}
\details{
If there are more years of deaths than years of population, then the last
year of available population data is used to compute the rates for
subsequent years.
}
\examples{

nsw.M <- make.demogdata(state="NSW",cut="M")
plot(nsw.M, "female")

# Show smoothing
test <- make.demogdata(state="NSW", cut="M", sex="female", smooth=FALSE)
test.sm <- smooth.demogdata(test)
plot(test, year=2006, type="p", pch=1)
lines(test.sm, year=2006, col="red")

}
\references{
Choi, C., Hyndman, R.J., Smith, L., and Zhao, K. (2010) \emph{An
enhanced mortality database for estimating indigenous life expectancy}.
Report for Australian Institute of Health and Welfare.
}
\author{
Rob J Hyndman <Rob.Hyndman@monash.edu>
}
