% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simcut.R
\name{simulate.demogdata}
\alias{simulate.demogdata}
\title{Simulate a new demogdata object based on an existing demogdata object using
a parametric bootstrap.}
\usage{
\method{simulate}{demogdata}(object, nsim = 1, seed = NULL,
  sigma = 0.0295, ...)
}
\arguments{
\item{object}{A demogdata object that will be used as the basis for the
simulation.}

\item{nsim}{Number of simulated populations returned. This argument is
ignored.}

\item{seed}{Either NULL or an integer that will be used in a call to
set.seed before simulation. The default, NULL will not change the random
generator state.}

\item{sigma}{PES sigma estimate.}

\item{...}{Other arguments are ignored.}
}
\value{
A demogdata object of the same structure as \code{object}.
}
\description{
Populations are simulated using a normal distribution with standard
deviation based on PES estimates. Deaths are estimated using a Poisson
distribution. This function assumes that the rates for the demogdata object
used as an argument are smoothed. When the simulated deaths are equal to
zero (so the simulated rate is then either zero or undefined), the simulated
rate is set to be equal to the smoothed rate contained in \code{object}.
Otherwise, the simulated rate is set to be the simulated deaths divided by
the simulated population.
}
\examples{

plot(nt$N,year=2003)
points(simulate(nt$N),year=2003)

# Example of infant.mortality CI for NT males
nsim <- 1000
infant.mort <- numeric(nsim)
ntmale <- make.demogdata(state="NT",cut="N",sex="male",aveyear=TRUE)
for(i in 1:nsim)
{
  sim <- simulate(ntmale)
  infant.mort[i] <- sim$rate$male[1,1]
}
# 95\% CI for infant mortality rate per 100,000 boys aged 0
# (not the same as per 100,000 births)
quantile(infant.mort,c(0.025,0.975))*100000

# Example of total mortality rate for WA
nsim <- 1000
TMR <- matrix(NA,nsim,3)
wa <- make.demogdata(state="WA", cut="N", aveyear=TRUE)
colnames(TMR) <- c("male","female","total")
for(i in 1:nsim)
{
  sim <- simulate(wa)
  TMR[i,1] <- sum(sim$rate$male * sim$pop$male) / sum(sim$pop$male)
  TMR[i,2] <- sum(sim$rate$female * sim$pop$female) / sum(sim$pop$female)
  TMR[i,3] <- sum(sim$rate$total * sim$pop$total) / sum(sim$pop$total)
}
# 95\% CI for total mortality rate per 100,000 people
# (not age-standardized)
TMR.ci <- apply(TMR,2,quantile,prob=c(0.025,0.975))*100000

}
\references{
Choi, C., Hyndman, R.J., Smith, L., and Zhao, K. (2010) \emph{An
enhanced mortality database for estimating indigenous life expectancy}.
Report for Australian Institute of Health and Welfare.
}
\author{
Rob J Hyndman <Rob.Hyndman@monash.edu>
}
