% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indmortality-package.R
\docType{data}
\name{aus}
\alias{aus}
\alias{nsw}
\alias{vic}
\alias{qld}
\alias{nt}
\alias{sa}
\alias{wa}
\title{Lists of demogdata objects for all Australian states, 2001-2007}
\description{
Objects: \code{aus}, \code{nsw}, \code{vic}, \code{qld}, \code{wa},
\code{sa} and \code{nt}.
}
\details{
These are lists of demogdata objects. Each state contains four demogdata
objects, corresponding to the M, H, A and N cuts of the indigenous deaths
data. For information about demogdata objects, see
\code{\link[demography]{demogdata}}.

Demogdata objects constructed using \code{\link{make.demogdata}} using
2001-2007 data from \code{\link{ideaths}} and \code{\link{ipop}}.
}
\examples{

plot(nsw$M,'male')

}
\keyword{datasets}
