% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtable.R
\name{deathstable}
\alias{deathstable}
\title{Function to make deaths tables for subsets of data.}
\usage{
deathstable(state = c("AUS", "ACT", "NSW", "NT", "QLD", "SA", "TAS", "VIC",
  "WA"), sex = c("female", "male", "total"), linked = c(NA, TRUE, FALSE),
  classified = c(NA, "N", "Y", "U"), year = NULL, aveyear = FALSE,
  upper.age = 100, cumulative = TRUE, deaths = indmortality::ideaths)
}
\arguments{
\item{state}{If "ALL", the function will return all deaths with state as a
dimension of the array. Otherwise the function will return all deaths from
the state indicated but with state not a dimension of the array.}

\item{sex}{If "ALL", the function will return all deaths with sex as a
dimension of the array. Otherwise the function will return all deaths from
the sex indicated but with sex not a dimension of the array.}

\item{year}{Which years to be returned. Default is all available years.
years is always a dimension of the returned array.}

\item{aveyear}{If TRUE, the deaths are averaged across years.}

\item{upper.age}{Deaths are cumulated for higher ages into this upper age
group.}

\item{cumulative}{Indicates if the deaths are to be cumulated up to the
specified "cut", or whether the marginal deaths associated with that cut are
to be returned.}

\item{deaths}{Matrix containing all deaths with indigenous classification in
Australia from 2001-2007 in the same format as \code{\link{deaths}}.}

\item{cut}{If "ALL", the function will return all deaths with state as a
dimension of the array. Otherwise the function will return all deaths from
the state indicated but with state not a dimension of the array.}
}
\value{
A numerical array with death counts in each cell. Dimensions of the
array are age, year and then (depending on the arguments) state,sex,cut.
}
\description{
Returns array with ages and years and (depending on the arguments) states,
sex and cut. Default arguments will return the complete array. Use the
argument to specify subsets of the array. Uses all available years unless
aveyear=TRUE, in which case it will return average deaths across years.
}
\examples{

tab1 <- deathstable(sex="female")
tab2 <- deathstable(state="VIC", cut="H", aveyear=TRUE)
tab3 <- deathstable()

}
\references{
Choi, C., Hyndman, R.J., Smith, L., and Zhao, K. (2010) \emph{An
enhanced mortality database for estimating indigenous life expectancy}.
Report for Australian Institute of Health and Welfare.
}
\author{
Rob J Hyndman <Rob.Hyndman@monash.edu>
}
