% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indmortality-package.R
\docType{data}
\name{ideaths}
\alias{ideaths}
\title{Database of all deaths with indigenous classification in Australia from
2001-2007.}
\format{A data frame with 16625 observations on the following six variables:
\describe{ \item{list("State")}{a factor with levels \code{ACT}, \code{NSW},
\code{NT}, \code{QLD}, \code{SA}, \code{TAS}, \code{VIC}, \code{WA}.}
\item{list("Year")}{a numeric vector indicating year of registration.}
\item{list("Deathyear")}{a numeric vector indicating year of death.}
\item{list("Sex")}{a factor with levels \code{F} and \code{M}.}
\item{list("Age")}{a numeric vector indicating age at death.}
\item{list("Cut")}{a factor with levels \code{M}, \code{H}, \code{A} and
\code{N} indicating the database where the death was first recorded
(treating the levels in order).} }}
\description{
Database constructed by linking National Mortality Database, hospital
records, residential aged care records and neonatal records. The
\code{ideaths} matrix was formed using the \code{\link{read.deaths}} command
using the data set available on 22 November 2011.
}
\examples{

summary(ideaths)

}
\references{
Choi, C., Hyndman, R.J., Smith, L., and Zhao, K. (2010) \emph{An
enhanced mortality database for estimating indigenous life expectancy}.
Report for Australian Institute of Health and Welfare.
}
\keyword{datasets}
