% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indmortality-package.R
\docType{package}
\name{indmortality-package}
\alias{indmortality-package}
\alias{indmortality}
\alias{indmortality2}
\title{Australian indigenous mortality rates and life expectancies}
\description{
Package prepared for the Australian Institute of Health and Welfare to
compute life expectancy and other mortality-related information for the
Australian indigenous population using simulation and nonparametric
smoothing.
}
\examples{

nsw <- make.demogdata(state="NSW", population="backcast")
plot(nsw)
life.expectancy.mean(nsw)
life.expectancy.ci(nsw)

# Complete calculation for Australia, 2001-2005.
aus1 <- make.demogdata(state="AUS", year=2001:2005, aveyear=TRUE,
   population="cohort")
aus2 <- make.demogdata(state="AUS", year=2001:2005, aveyear=TRUE,
   population="interpolated")
aus3 <- make.demogdata(state="AUS", year=2001:2005, aveyear=TRUE,
   population="backcast")
life.expectancy.mean(aus1)
life.expectancy.ci(aus1)
life.expectancy.mean(aus2)
life.expectancy.ci(aus2)
life.expectancy.mean(aus3)
life.expectancy.ci(aus3)

}
\references{
Choi, C., Hyndman, R.J., Smith, L., and Zhao, K. (2010) \emph{An
enhanced mortality database for estimating indigenous life expectancy}.
Report for Australian Institute of Health and Welfare.
}
\seealso{
\code{\link[demography:demography-package]{demography}}
}
\author{
Rob J Hyndman <Rob.Hyndman@monash.edu>
}
\keyword{package}
