% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indmortality-package.R
\docType{data}
\name{ipop}
\alias{ipop}
\title{Populations for all Australian states, 2001-2006}
\source{
ABS Cat 3238.0
}
\description{
This is a list of lists of matrices. The first level is the Australian
states and the second level contains sex groups. Each matrix is of size
101x6 with single ages in each row and calendar years in columns. For
example, ipop$NSW$male contains the estimated indigenous male population for
NSW in 2001-2006.
}
\details{
Population estimates are based on ABS 3238.0 which gives five year age
groups up to 85+. These are distributed according to earlier single-year-age
estimates of the state indigenous populations, except for the ACT for which
the five-year age groups are distributed according to the NSW
single-year-age estimates of the indigenous population. The population for
85+ in each state is distributed according to the general Australian age
distribution for that age group.
}
\examples{

plot(0:100,ipop$NT$male[,"2003"],type="l",xlab="Age",ylab="Population",
  main="Northern Territory 2003",col="blue")
lines(0:100,ipop$NT$female[,"2003"],col="red")
legend("topright",lty=1,col=c("blue","red"),legend=c("male","female"))

}
\keyword{datasets}
