% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.demogdata.R
\name{make.demogdata}
\alias{make.demogdata}
\title{Make demogdata objects for particular states and years.}
\usage{
make.demogdata(state = c("AUS", "ACT", "NSW", "NT", "QLD", "SA", "TAS", "VIC",
  "WA"), year = 2001:2010, aveyear = FALSE, upper.age = 100,
  smooth = TRUE, population = c("cohort", "interpolated", "backcast"))
}
\arguments{
\item{state}{Character code indicating state or territory of Australia, or
"AUS" indicating the whole of Australia.}

\item{year}{Years to include in the demogdata object. The default is to
include all available years from the \link{ideaths} database.}

\item{aveyear}{Logical value indicating whether deaths and population numbers
should be averaged across years. Default is FALSE.}

\item{upper.age}{Upper age group, by default set to 100+.}

\item{smooth}{Logical value indicating whether mortality rates should be
smoothed using penalized regression splines. Default is TRUE.}

\item{population}{Character code indicating which estimated residential
indigenous population should be used. \code{"cohort"} means use the ERP
obtained by linearly interpolating the 2001, 2006 and 2011 census figures
along cohorts. \code{"interpolated"} means use the ERP obtained by linearly
interpolating the 2001, 2006 and 2011 census figures along ages.
\code{"backcast"} means use the ABS ERP based on 2011 census values.}
}
\value{
A demogdata object containing mortality rates for males, females and total.
See \link[demography]{demogdata} for more information about demogdata objects.
}
\description{
The function \code{make.demogdata} will return a demogdata object
for the given state and years.
}
\examples{

nsw <- make.demogdata(state="NSW")
plot(nsw, "female")

# Show smoothing
test <- make.demogdata(state="NSW", smooth=FALSE)
test.sm <- demography::smooth.demogdata(test)
plot(test, series='female', year=2006, type="p", pch=1)
lines(test.sm, series='female', year=2006, col="red")

}
\references{
Choi, C., Hyndman, R.J., Smith, L., and Zhao, K. (2010) \emph{An
enhanced mortality database for estimating indigenous life expectancy}.
Report for Australian Institute of Health and Welfare.
}
\author{
Rob J Hyndman <Rob.Hyndman@monash.edu>
}
